Implement template approach
previously, we tried to make a list of elements in sequence and see which fit.

This approach goes the other way. We first generate a list of templates e.g. *____*_____ with * marking black squares. We then combine this with the constraints imposed by the regions. E.g. the first row is "SSSDSSDSDSS" S: means same as previous, D means different to previous. These two rows are combined together to yield a final constraint set. Either != 0, != previous, == previous or *. Then using a backtracking approach, we see which numbers can satisfy those constraints. At the occurence of a * or reaching the end, the backtracking algorithm checks that the latest number generated e.g. ___*225* satisfies the clue constraint. If not this path is pruned. Otherwise continue searching.

This approach has been applied to row 1 already. Next we'll apply this to the other rows. This should create solutions for each row. Then we will try to stack the rows together so that each row solution is compatable with each other