Fixed Issue with template generation in main.rs
We forgot to add a string in an array. This meant we were missing template cases.

This caused issues when trying to concatenate

We also started creating concatenate.rs

We tried to be fancy with a "recursive"-ish solver to backtrack using some nested while loops. It doesn't work right now.

What instead works is just pairwise checking for compatability.

KISS.

We're going to stick with the simple pairwise method moving forward. If we want to slightly automate the pairs, we can just turn the pairs into edges of a graph and then Djkstra it